fix(frontend): share view test harness
