+++ frontend/src/views/events/Apply.vue
@@ -230,8 +230,9 @@ export default {
         })
       }).catch((err) => {
         this.isSaving = false
+        const status = err.response && err.response.status
 
-        if (err.response.status === 422) { // validation errors
+        if (status === 422) { // validation errors
           this.errors = err.response.data.errors
           return this.$root.showError('Some of the application data is invalid.')
         }
@@ -284,13 +285,17 @@ export default {
         this.isLoading = false
       }).catch((err) => {
         this.isLoading = false
+        const status = err.response && err.response.status
+
         // if there's no application, just ignore, otherwise re-throw the error
-        if (err.response.status !== 404) {
+        if (status !== 404) {
           throw err
         }
       })
     }).catch((err) => {
-      if (err.response.status === 404) {
+      const status = err.response && err.response.status
+
+      if (status === 404) {
         this.$root.showError('Event is not found')
       } else {
         this.$root.showError('Some error happened', err)
